requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
WTForms>=3.1.0
marshmallow>=3.20.0
flask-marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0
email-validator>=2.0.0